+++ ReactCommon/fabric/components/image/BUCK
@@ -58,7 +58,7 @@ rn_xplat_cxx_library(
         react_native_xplat_target("fabric/core:core"),
         react_native_xplat_target("fabric/graphics:graphics"),
         react_native_xplat_target("fabric/imagemanager:imagemanager"),
-        react_native_xplat_target("fabric/view:view"),
+        react_native_xplat_target("fabric/components/view:view"),
     ],
 )
