+++ ReactCommon/fabric/components/image/ImageProps.h
@@ -5,9 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+#include <fabric/components/view/ViewProps.h>
 #include <fabric/graphics/Color.h>
 #include <fabric/imagemanager/primitives.h>
-#include <fabric/view/ViewProps.h>
 
 namespace facebook {
 namespace react {
