+++ ReactCommon/fabric/components/switch/SwitchShadowNode.h
@@ -9,7 +9,7 @@
 
 #include <fabric/components/switch/SwitchEventEmitter.h>
 #include <fabric/components/switch/SwitchProps.h>
-#include <fabric/view/ConcreteViewShadowNode.h>
+#include <fabric/components/view/ConcreteViewShadowNode.h>
 
 namespace facebook {
 namespace react {
