+++ ReactCommon/fabric/components/text/BUCK
@@ -66,7 +66,7 @@ rn_xplat_cxx_library(
         react_native_xplat_target("fabric/debug:debug"),
         react_native_xplat_target("fabric/graphics:graphics"),
         react_native_xplat_target("fabric/textlayoutmanager:textlayoutmanager"),
-        react_native_xplat_target("fabric/view:view"),
+        react_native_xplat_target("fabric/components/view:view"),
     ],
 )
