+++ ReactCommon/fabric/components/view/BUCK
@@ -31,7 +31,7 @@ rn_xplat_cxx_library(
             ("root", "*.h"),
             ("yoga", "*.h"),
         ],
-        prefix = "fabric/view",
+        prefix = "fabric/components/view",
     ),
     compiler_flags = [
         "-fexceptions",
