+++ ReactCommon/fabric/components/view/accessibility/AccessibilityProps.h
@@ -7,10 +7,10 @@
 
 #pragma once
 
+#include <fabric/components/view/AccessibilityPrimitives.h>
 #include <fabric/core/Props.h>
 #include <fabric/core/ReactPrimitives.h>
 #include <fabric/debug/DebugStringConvertible.h>
-#include <fabric/view/AccessibilityPrimitives.h>
 
 namespace facebook {
 namespace react {
