+++ ReactCommon/fabric/components/view/propsConversions.h
@@ -7,8 +7,8 @@
 
 #pragma once
 
+#include <fabric/components/view/conversions.h>
 #include <fabric/core/propsConversions.h>
-#include <fabric/view/conversions.h>
 
 namespace facebook {
 namespace react {
