+++ ReactCommon/fabric/components/view/yoga/YogaStylableProps.cpp
@@ -7,9 +7,9 @@
 
 #include "YogaStylableProps.h"
 
-#include <fabric/view/conversions.h>
+#include <fabric/components/view/conversions.h>
 #include <fabric/core/propsConversions.h>
-#include <fabric/view/propsConversions.h>
+#include <fabric/components/view/propsConversions.h>
 #include <fabric/debug/debugStringConvertibleUtils.h>
 #include <yoga/YGNode.h>
 #include <yoga/Yoga.h>
