+++ ReactCommon/fabric/uimanager/Scheduler.cpp
@@ -9,12 +9,12 @@
 #include <fabric/components/text/ParagraphComponentDescriptor.h>
 #include <fabric/components/text/RawTextComponentDescriptor.h>
 #include <fabric/components/text/TextComponentDescriptor.h>
+#include <fabric/components/view/ViewComponentDescriptor.h>
+#include <fabric/components/view/ViewProps.h>
+#include <fabric/components/view/ViewShadowNode.h>
 #include <fabric/core/LayoutContext.h>
 #include <fabric/uimanager/ComponentDescriptorRegistry.h>
 #include <fabric/uimanager/FabricUIManager.h>
-#include <fabric/view/ViewComponentDescriptor.h>
-#include <fabric/view/ViewProps.h>
-#include <fabric/view/ViewShadowNode.h>
 
 #include "ComponentDescriptorFactory.h"
 #include "Differentiator.h"
