+++ ReactCommon/fabric/uimanager/Scheduler.h
@@ -7,6 +7,8 @@
 
 #include <memory>
 
+#include <fabric/components/view/ViewShadowNode.h>
+#include <fabric/components/view/RootShadowNode.h>
 #include <fabric/core/ComponentDescriptor.h>
 #include <fabric/core/LayoutConstraints.h>
 #include <fabric/uimanager/ContextContainer.h>
@@ -15,8 +17,6 @@
 #include <fabric/uimanager/UIManagerDelegate.h>
 #include <fabric/uimanager/ShadowTree.h>
 #include <fabric/uimanager/ShadowTreeDelegate.h>
-#include <fabric/view/ViewShadowNode.h>
-#include <fabric/view/RootShadowNode.h>
 
 namespace facebook {
 namespace react {
