+++ ReactCommon/fabric/uimanager/ShadowTree.h
@@ -8,11 +8,11 @@
 #include <memory>
 #include <mutex>
 
+#include <fabric/components/view/RootShadowNode.h>
 #include <fabric/core/LayoutConstraints.h>
 #include <fabric/core/ReactPrimitives.h>
 #include <fabric/core/ShadowNode.h>
 #include <fabric/uimanager/ShadowTreeDelegate.h>
-#include <fabric/view/RootShadowNode.h>
 
 namespace facebook {
 namespace react {
